- updated Milepost class to use BigDecimal instead of double due to data loss
- associated updates for BigDecimal update
- switched region path to use a delta of `node-LL`
- added calculation for offset rather than absolute lat/lon
- updated path scale and type to reflect offset

Related work items: #5274

+++ cv-data-controller/src/test/java/com/trihydro/cvdatacontroller/controller/MilepostControllerTest.java
@@ -251,8 +251,8 @@ public class MilepostControllerTest extends TestBase<MilepostController> {
         verify(mockStatement).executeQuery(statementStr);
         verify(mockRs).getString("COMMON_NAME");
         verify(mockRs).getDouble("MILEPOST");
-        verify(mockRs).getDouble("LATITUDE");
-        verify(mockRs).getDouble("LONGITUDE");
+        verify(mockRs).getBigDecimal("LATITUDE");
+        verify(mockRs).getBigDecimal("LONGITUDE");
         // verify(mockRs).getDouble("BEARING");
         verify(mockStatement).close();
         verify(mockConnection).close();
@@ -278,8 +278,8 @@ public class MilepostControllerTest extends TestBase<MilepostController> {
         verify(mockStatement).executeQuery(statementStr);
         verify(mockRs).getString("COMMON_NAME");
         verify(mockRs).getDouble("MILEPOST");
-        verify(mockRs).getDouble("LATITUDE");
-        verify(mockRs).getDouble("LONGITUDE");
+        verify(mockRs).getBigDecimal("LATITUDE");
+        verify(mockRs).getBigDecimal("LONGITUDE");
         // verify(mockRs).getDouble("BEARING");
         verify(mockStatement).close();
         verify(mockConnection).close();
